requirements.txt
requests
keyboard
riotwatcher
Django==2.2.2
zappa==0.48.2
django-cors-headers==2.4.0
django-cors-middleware==1.3.1
djangorestframework==3.9.4
mysqlclient
pynamodb==3.4.1
drf-yasg==1.16.0
drf-writable-nested==0.5.1